Tests/LibWeb/Text/input/DOM/getElementsName-empty-string.html

<!DOCTYPE html>
<script src="../include.js"></script>
<div name=""></div>
<script>
    test(() => {
        println(`document.getElementsByName("").length: ${document.getElementsByName("").length}`);
    });
</script>
