Tests/LibWeb/Text/expected/DOM/getElementsName-empty-string.txt

document.getElementsByName("").length: 0
